requirements.txt
pytest~=8.4.1
allure-python-commons~=2.14.3
requests~=2.32.4
Faker~=37.4.2